apidevtools/template/template/api/requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
cchardet==2.1.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.90.1
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
loguru==0.6.0
multidict==6.0.4
numpy==1.24.1
Pillow==9.4.0
pycares==4.3.0
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.4
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.5
python-string-utils==1.0.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
telegraph==2.2.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
win32-setctime==1.1.0
yarl==1.8.2
